requirements.txt
# Core dependencies
python-dotenv==1.2.1
openai==2.7.1
groq==0.33.0

# LangChain ecosystem
langchain==1.0.3
langchain-core==1.0.3
langchain-community==0.4.1
langchain-openai==1.0.2

# Embeddings and vector storage
faiss-cpu==1.12.0
sentence-transformers==5.1.2
huggingface-hub==0.36.0

# PDF text extraction
pdfplumber==0.11.7
pdfminer.six==20250506

# Utilities
numpy==2.3.4
scikit-learn==1.7.2
tqdm==4.67.1
requests==2.32.5
python-dotenv==1.2.1
Streamlit==1.51.0

# Optional (if used by LangChain pipelines)
pydantic==2.12.4
pydantic-core==2.41.5
tenacity==9.1.2

